+++ src/test/java/com/atlas/tool/impl/ListToolParameterPassThroughContractTest.java
@@ -61,6 +61,8 @@ class ListToolParameterPassThroughContractTest {
         assertPassThrough(SlurmNodeListTool::new, "/api/100002/slurm-node");
         assertPassThrough(SlurmClusterListTool::new, "/api/100002/bcm/slurm-cluster");
         assertPassThrough(UploadStatusListTool::new, "/api/100002/download/status");
+        assertPassThrough(ResourceUsageListTool::new, "/api/100002/resource");
+        assertPassThrough(QuotaMyListTool::new, "/api/100002/quota/my");
     }
 
     @Test
@@ -124,6 +126,8 @@ class ListToolParameterPassThroughContractTest {
         assertInvalidPagination(DownloadTaskListTool::new, "/api/100002/download", "limit", 1.5D, "TYPE_MISMATCH");
         assertInvalidPagination(SlurmClusterListTool::new, "/api/100002/bcm/slurm-cluster", "page", "abc", "TYPE_MISMATCH");
         assertInvalidPagination(UploadStatusListTool::new, "/api/100002/download/status", "limit", "0", "VALUE_OUT_OF_RANGE");
+        assertInvalidPagination(ResourceUsageListTool::new, "/api/100002/resource", "page", "0", "VALUE_OUT_OF_RANGE");
+        assertInvalidPagination(QuotaMyListTool::new, "/api/100002/quota/my", "limit", 1.5D, "TYPE_MISMATCH");
     }
 
     private void assertPassThrough(ToolFactory factory, String expectedPath) {