+++ src/test/java/com/atlas/tool/impl/ListToolParameterSpecContractTest.java
@@ -50,6 +50,8 @@ class ListToolParameterSpecContractTest {
         assertListQueryContract(new SlurmNodeListTool(null), "slurm_node_list");
         assertListQueryContract(new SlurmClusterListTool(null), "slurm_cluster_list");
         assertListQueryContract(new UploadStatusListTool(null), "upload_status_list");
+        assertListQueryContract(new ResourceUsageListTool(null), "resource_usage_list");
+        assertListQueryContract(new QuotaMyListTool(null), "quota_my_list");
     }
 
     /**